requirements/requirements.txt
wheel
bar
matplotlib==3.3.4
imutils
rtree
dill
pandas
seaborn
opencv-python
pyyaml
tqdm
pyzmq
faiss-cpu
pycocotools
jsonpickle
-e ..